feat: separate search to client and server
