+++ src/app/(server)/_components/ui/search-input.tsx
@@ -1,4 +1,3 @@
-'use client';
 import { cn } from '@/app/(server)/_lib/cn';
 
 type SearchInputProps = {
