+++ src/app/page.tsx
@@ -6,13 +6,15 @@ import useSWR from 'swr';
 
 import { Pagination } from '@/app/_components/pagination';
 import { RefreshButton } from '@/app/_components/refresh-button';
-import Search from '@/app/_components/search';
 
 import { useCharacters } from './(client)/_hooks/use-characters';
+import useLocalStorage from './(client)/_hooks/use-local-storage';
 import Results from './(server)/_components/data/results';
+import Search from './(server)/_components/search';
 import { Loader } from './(server)/_components/ui/loader';
 import { NotSearchFound } from './(server)/_components/ui/not-found-search';
 import { cn } from './(server)/_lib/cn';
+import { SEARCH_TERM_KEY } from './(server)/_lib/constants';
 import { fetchCharacters } from './(server)/_services/fetch-characters';
 import type { ApiResponse } from './_types/api';
 
@@ -24,17 +26,18 @@ export default function Home() {
     ['characters', currentSearch, currentPage],
     () => fetchCharacters({ name: currentSearch, page: currentPage }),
   );
-
+  const [searchTerm, setSearchTerm] = useLocalStorage(SEARCH_TERM_KEY, '');
   const hasDetails = pathname.includes('/characters/');
   const handleSearchWithReset = (searchTerm: string) => {
     handleSearch(searchTerm);
+    setSearchTerm(searchTerm);
     if (hasDetails) router.push('/');
   };
 
   return (
     <div className="mx-auto my-20 flex w-full flex-col items-center px-8 py-4">
       <h1 className="text-2xl font-bold">Rick & Morty</h1>
-      <Search onSearch={handleSearchWithReset} />
+      <Search onSearch={handleSearchWithReset} initialSearchTerm={searchTerm} />
       <Suspense fallback={<Loader />}>{isLoading && <Loader />}</Suspense>
 
       {error ? (
